[Add, Feat, Refactor] StoresViewModel에 있는 로직을 ExploreViewModel으로 이전 - ExploreView에서만 사용하는 메서드분리, KingFisher로 이미지 캐싱, ExploreView 코드 재사용성 높이는 방향으로 리팩토링
