+++ GukbapMinister/GukbapMinister/ViewModels/Store/StoresViewModel.swift
@@ -18,131 +18,12 @@ final class StoresViewModel: ObservableObject {
     @Published var stores: [Store] = []
     @Published var storeTitleImage: [String : UIImage] = [:]
     
-    @Published var storesStar: [Store] = []
-    @Published var storeTitleImageStar: [String : UIImage] = [:]
-    
     @Published var countRan = 0
 
-    @Published var storesHits: [Store] = []
-    @Published var storeTitleImageHits: [String : UIImage] = [:]
-    
     private var database = Firestore.firestore()
     private var storage = Storage.storage()
     private var listenerRegistration: ListenerRegistration?
-    
-    
-    
-    func fetchStarStores() {
-            let ref = database.collection("Store").order(by: "countingStar", descending: true)
-            
-            ref.getDocuments { snapShot, error in
-                
-                self.storesStar.removeAll()
-                
-                if let snapShot {
-                    for document in snapShot.documents {
-                        let id: String = document.documentID
-                        let docData = document.data()
-                        
-                        let storeName: String = docData["storeName"] as? String ?? ""
-                        let storeAddress: String = docData["storeAddress"] as? String ?? ""
-                        let coordinate: GeoPoint = docData["coordinate"] as? GeoPoint ?? GeoPoint(latitude: 0.0, longitude: 0.0)
-                        let storeImages: [String] = docData["storeImages"] as? [String] ?? []
-                        let menu: [String : String] = docData["menu"] as? [String : String] ?? ["":""]
-                        let description: String = docData["description"] as? String ?? ""
-                        let countingStar: Double = docData["countingStar"] as? Double ?? 0
-                        let foodType : [String] = docData["foodType"]  as? [String] ?? []
-                        let likes : Int = docData["likes"]  as? Int ?? 0
-                        let hits : Int = docData["hits"] as? Int ?? 0
-
-
-                        for imageName in storeImages {
-                            self.fetchImagesStar(storeId: storeName, imageName: imageName)
-                        }
-
-                        let store: Store = Store(id: id,
-                                                 storeName: storeName,
-                                                 storeAddress: storeAddress,
-                                                 coordinate: coordinate,
-                                                 storeImages: storeImages,
-                                                 menu: menu,
-                                                 description: description,
-                                                 countingStar: countingStar,
-                                                 foodType: foodType,
-                                                 likes: likes,
-                                                 hits: hits)
-                         
-                        
-                        self.storesStar.append(store)
-                    }
-                }
-            }
-    }
-    
-    
-    
-    func fetchHitsStores() {
-            let ref = database.collection("Store").order(by: "hits", descending: true)
-
-            ref.getDocuments { snapShot, error in
-
-                self.storesHits.removeAll()
-
-                if let snapShot {
-                    for document in snapShot.documents {
-                        let id: String = document.documentID
-                        let docData = document.data()
-
-                        let storeName: String = docData["storeName"] as? String ?? ""
-                        let storeAddress: String = docData["storeAddress"] as? String ?? ""
-                        let coordinate: GeoPoint = docData["coordinate"] as? GeoPoint ?? GeoPoint(latitude: 0.0, longitude: 0.0)
-                        let storeImages: [String] = docData["storeImages"] as? [String] ?? []
-                        let menu: [String : String] = docData["menu"] as? [String : String] ?? ["":""]
-                        let description: String = docData["description"] as? String ?? ""
-                        let countingStar: Double = docData["countingStar"] as? Double ?? 0
-                        let foodType : [String] = docData["foodType"]  as? [String] ?? []
-                        let likes : Int = docData["likes"]  as? Int ?? 0
-                        let hits : Int = docData["hits"] as? Int ?? 0
-
-                        for imageName in storeImages {
-                            self.fetchImagesHits(storeId: storeName, imageName: imageName)
-                        }
-
-                        let store: Store = Store(id: id,
-                                                 storeName: storeName,
-                                                 storeAddress: storeAddress,
-                                                 coordinate: coordinate,
-                                                 storeImages: storeImages,
-                                                 menu: menu,
-                                                 description: description,
-                                                 countingStar: countingStar,
-                                                 foodType: foodType,
-                                                 likes: likes,
-                                                 hits: hits)
-
-                        self.storesHits.append(store)
-                    }
-                }
-            }
-    }
-
-
-    // 조회수를 increase시키는 메서드 (onAppear시 동작)
-    func increaseHits(store : Store) {
-        var updateStoreHits = store
-        updateStoreHits.hits += 1
-
-        if let documentId = store.id {
-            do {
-                try database.collection("Store")
-                    .document(documentId)
-                    .updateData(["hits" : updateStoreHits.hits])
-            }
-            catch {
-                print(error)
-            }
-        }
-    }
+ 
     
     //Store정보 구독취소
     //Store정보가 필요한 뷰에서
@@ -155,6 +36,7 @@ final class StoresViewModel: ObservableObject {
     }
     
     // 생성 시점 이슈로 인해 뷰모델에서 난수를 생성
+    //FIXME: collectionViewModel로 옮길것!!!
     func getRandomNumber() {
         database.collection("Store").getDocuments()
         {
@@ -182,9 +64,6 @@ final class StoresViewModel: ObservableObject {
     //.onAppear { viewModel.subscribeStores() } 하면 실행됨
     func subscribeStores() {
         if listenerRegistration == nil {
-
-            
-            
             listenerRegistration =  database.collection("Store")
                 .addSnapshotListener { (querySnapshot, error) in
                     guard let documents = querySnapshot?.documents else {
@@ -194,7 +73,6 @@ final class StoresViewModel: ObservableObject {
                     
                     //FirebaseFireStoreSwift 를 써서 @Document 프로퍼티를 썼더니 가능
                     self.stores = documents.compactMap { queryDocumentSnapshot in
-
                         let result = Result { try queryDocumentSnapshot.data(as: Store.self) }
 
                         switch result {
@@ -220,25 +98,6 @@ final class StoresViewModel: ObservableObject {
         }
     }
     
-    
-    
-    
-    // MARK: - Storage에서 이미지 다운로드
-//    private func fetchImages(storeId: String, imageName: String) {
-//        let ref = storage.reference().child("storeImages/\(storeId)/\(imageName)")
-//
-//        // Download in memory with a maximum allowed size of 1MB (1 * 1024 * 1024 bytes)
-//        ref.getData(maxSize: 15 * 1024 * 1024) { [self] data, error in
-//            if let error = error {
-//                print("error while downloading image\n\(error.localizedDescription)")
-//                return
-//            } else {
-//                let image = UIImage(data: data!)
-//                self.storeTitleImage[imageName] = image
-//
-//            }
-//        }
-//    }
     func fetchImages(storeId: String, imageName: String) async throws -> UIImage {
         let ref = storage.reference().child("storeImages/\(storeId)/\(imageName)")
 
@@ -249,39 +108,7 @@ final class StoresViewModel: ObservableObject {
         
         return image!
     }
-    
-    
-    func fetchImagesStar(storeId: String, imageName: String) {
-        let ref = storage.reference().child("storeImages/\(storeId)/\(imageName)")
-        
-        // Download in memory with a maximum allowed size of 1MB (1 * 1024 * 1024 bytes)
-        ref.getData(maxSize: 15 * 1024 * 1024) { [self] data, error in
-            if let error = error {
-                print("error while downloading image\n\(error.localizedDescription)")
-                return
-            } else {
-                let image = UIImage(data: data!)
-                self.storeTitleImageStar[imageName] = image
-         
-            }
-        }
-    }
-    
-    func fetchImagesHits(storeId: String, imageName: String) {
-        let ref = storage.reference().child("storeImages/\(storeId)/\(imageName)")
 
-        // Download in memory with a maximum allowed size of 1MB (1 * 1024 * 1024 bytes)
-        ref.getData(maxSize: 15 * 1024 * 1024) { [self] data, error in
-            if let error = error {
-                print("error while downloading image\n\(error.localizedDescription)")
-                return
-            } else {
-                let image = UIImage(data: data!)
-                self.storeTitleImageHits[imageName] = image
-
-            }
-        }
-    }
     
 //    // MARK: 사용자 찜하기 데이터 수정 Method
 //    func fetchFavorite(isFavorited: Bool, user: User, storeId: String) async {
