+++ GukbapMinister/GukbapMinister/Views/MapView/MapView.swift
@@ -4,13 +4,12 @@ import CoreLocation
 import CoreLocationUI
 
 struct MapView: View {
-
     @Environment(\.colorScheme) var scheme
-    // Command + F -> replace: changes searched word in the file
-    @EnvironmentObject var storesViewModel: StoresViewModel
     @StateObject var mapViewModel = MapViewModel(storeLocations: [])
     @StateObject var locationManager = LocationManager()
+    
     @EnvironmentObject var userViewModel : UserViewModel
+    @EnvironmentObject var storesViewModel: StoresViewModel
     // 필터 버튼을 눌렀을 때 동작
     @State var isShowingFilterModal: Bool = false
     @State private var isShowingSelectedStore: Bool = false
@@ -63,14 +62,12 @@ struct MapView: View {
             
         }
         .onAppear {
-            Task{
-                mapViewModel.storeLocations = storesViewModel.stores
-                DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-                    mapViewModel.setStoreLocationAnnotations()
+            Task {
+                DispatchQueue.main.asyncAfter(deadline:.now() + 0.5) {
+                    mapViewModel.storeLocations = storesViewModel.stores
                 }
             }
         }
-        
     }
     
 }
